main/requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1.2
django-cors-headers==4.5.0
django-debug-toolbar==4.4.6
django-prometheus==2.3.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.1
prometheus_client==0.21.0
psycopg2-binary==2.9.9
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
uritemplate==4.1.1
